% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahp.R
\name{fahp}
\alias{fahp}
\title{FAHP}
\usage{
fahp(zipdata, y, q = 2, maxit = 300, constraint = 5, para = 0)
}
\arguments{
\item{zipdata}{Observation matrix, commonly characterized by zero inflation.}

\item{y}{A response vector with dimension equal to the number of rows in zipdata.}

\item{q}{The number of latent factors. Default as 2.}

\item{maxit}{Maximum number of iterations within optim function, defaults to 300.}

\item{constraint}{Constraint constant. Default as 5.}

\item{para}{The number of cores used for parallel processing. Defaults to 0, indicating that parallel processing is not performed.}
}
\value{
\item{FA}{Factor score matrix.}
 \item{mu1}{The intercept vector for hurdle part}
 \item{mu2}{The intercept vector for positive count part}
 \item{t}{Number of iterations}
 \item{obj}{The loglikelihhod results (omitting the constant), which can be used for model selection}
}
\description{
Factor augmented hurdle Poisson inverse model. This method can conduct supervised dimensionality reduction on high dimensional count data containing a large number of zeros.
}
\examples{
n=50
p=20
x=matrix(rpois(n*p,lambda = 1),ncol = p)
y=rnorm(n)
fahp(x,y,q=1,maxit = 30,para = 10)
}
\author{
Zhijing Wang
}
